rest-api/requirements.txt
Flask
python-dotenv
flask-cors